Create Bond_with_subcells
This script calculates the bonds for each time step from the UMD file. It uses a structures of subcells to divide the working space in order to optimize the running time. The ration size of the space/sice of the cell can be specified with the option -n. The option -v allows to choose between different ways to compute : B is sequential, F is parallel with multiprocessing and the others use more ways to provide the same output.