Create msd_umd_par.py
This script fills the role of both msd_umd.py and msd_all_umd.py (depending on the option -a you use), while using parallelization and a C interface to compute each atom's msd. Use the option " -a species " to obtain the msd of each species averaged among its atoms. Use the option " -a atoms " to obtain the msd of each atom individually. Create the .so library beforehand from c_msd.c  and make sure it is located in the same repository than this .py script.